+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* 🎉 feat(global): global package added
 * 🎉 feat(fiber): fiber added as http and socker respose
 * ``README``, ``CHANGELOG`` and ``LICENSE`` added.
 * ``Translator`` package added.